+++ scripts/smart-dump.ts
@@ -5,8 +5,11 @@ import { execSync } from 'child_process';
 import { Command } from 'commander';
 import { readJsonSync, writeJsonSync } from 'fs-extra';
 import { uniqBy } from 'lodash';
+import moment from 'moment';
 
-import { traverse } from '../server/lib/export';
+import { traverse } from '../server/lib/import-export/export';
+import { restoreRows } from '../server/lib/import-export/import';
+import logger from '../server/lib/logger';
 import { md5 } from '../server/lib/utils';
 import models, { sequelize } from '../server/models';
 
@@ -16,18 +19,18 @@ const exec = cmd => {
   try {
     execSync(cmd, { stdio: 'inherit' });
   } catch (e) {
-    console.error(e);
+    logger.error(e);
   }
 };
 
 program.command('dump [recipe] [env]').action(async (recipe, env) => {
   if (!sequelize.config.username.includes('readonly')) {
-    console.error('Remote must be connected with read-only user!');
+    logger.error('Remote must be connected with read-only user!');
     process.exit(1);
   }
 
   if (!recipe || (recipe && !env)) {
-    console.log('Using default recipe...');
+    logger.info('Using default recipe...');
     recipe = './smart-dump/defaultRecipe.js';
   }
 
@@ -38,79 +41,71 @@ program.command('dump [recipe] [env]').action(async (recipe, env) => {
   const hash = md5(JSON.stringify({ entries, defaultDependencies, date })).slice(0, 5);
   const filename = `${date}.${hash}`;
   let docs = [];
-  console.time('>>> Dump');
+
+  let start = new Date();
+  logger.info('>>> Dumping...');
   for (const entry of entries) {
-    console.log(`\n>>> Traversing DB for entry ${entries.indexOf(entry) + 1}/${entries.length}...`);
+    logger.info(`>>> Traversing DB for entry ${entries.indexOf(entry) + 1}/${entries.length}...`);
     const newdocs = await traverse({ ...entry, defaultDependencies, parsed });
     docs.push(...newdocs);
   }
-  console.timeEnd('>>> Dump');
+  logger.info(`>>> Dumped! ${docs.length} records in ${moment(start).fromNow(true)}`);
 
-  console.log('\n>>> Deduplicating...');
+  logger.info('>>> Deduplicating...');
   docs = uniqBy(docs, r => `${r.model}.${r.id}`);
 
-  console.log('\n>>> Dumping JSON...');
+  start = new Date();
+  logger.info('>>> Writting JSON...');
   writeJsonSync(`dbdumps/${filename}.json`, docs, { spaces: 2 });
+  logger.info(`>>> Written! in ${moment(start).fromNow(true)}`);
 
-  console.log('\n>>> Dumping Schema...');
+  start = new Date();
+  logger.info('>>> Dumping Schema...');
   exec(`pg_dump -csOx $PG_URL > dbdumps/${filename}.schema.sql`);
+  logger.info(`>>> Schema Dumped! ${docs.length} records in ${moment(start).fromNow(true)}`);
 
-  console.log(`\n>>> Done! Dumped to dbdumps/${filename}.json`);
+  logger.info(`>>> Done! See dbdumps/${filename}.json and dbdumps/${filename}.schema.sql`);
   sequelize.close();
 });
 
 program.command('restore <file>').action(async file => {
   const database = process.env.PG_DATABASE;
   if (!database) {
-    console.error('PG_DATABASE is not set!');
+    logger.error('PG_DATABASE is not set!');
     process.exit(1);
   } else if (sequelize.config.database !== database) {
-    console.error(`Sequelize is not connected to target ${database}!`);
+    logger.error(`Sequelize is not connected to target ${database}!`);
     process.exit(1);
   }
 
-  console.log('\n>>> Recreating DB...');
+  let start = new Date();
+  logger.info('>>> Recreating DB...');
   exec(`dropdb ${database}`);
   exec(`createdb ${database}`);
   exec(`psql -h localhost -U opencollective ${database} < ${file.replace('.json', '.schema.sql')}`);
+  logger.info(`>>> DB Created! in ${moment(start).fromNow(true)}`);
 
   await sequelize.sync().catch(nop);
 
-  console.log(`\n>>> Reading file ${file}`);
+  logger.info(`>>> Reading file ${file}`);
   const docs = readJsonSync(file);
 
-  console.log('\n>>> Inserting Data...');
+  start = new Date();
+  logger.info('>>> Inserting Data...');
   const modelsArray: any[] = Object.values(models);
   for (const model of modelsArray) {
     const rows = docs.filter(d => d.model === model.name);
     if (rows.length > 0) {
-      console.log(`\t${model.name} (${rows.length} rows)`);
-      await sequelize
-        .transaction(async transaction => {
-          const tablename = model.getTableName();
-          await sequelize.query(`ALTER TABLE "${tablename}" DISABLE TRIGGER ALL;`, { transaction });
-          for (const row of rows) {
-            await model
-              .create(row, {
-                transaction,
-                validate: false,
-                hooks: false,
-                silent: true,
-                logging: false,
-                raw: false,
-                ignoreDuplicates: true,
-              })
-              .catch(console.error);
-          }
-          await sequelize.query(`ALTER TABLE "${tablename}" ENABLE TRIGGER ALL;`, { transaction });
-        })
-        .catch(e => {
-          console.error(e);
-        });
+      try {
+        await restoreRows(model, rows);
+      } catch (e) {
+        logger.error(e);
+      }
     }
   }
+  logger.info(`>>> Data inserted! in ${moment(start).fromNow(true)}`);
 
-  console.log('\n>>> Refreshing Materialized Views...');
+  logger.info('>>> Refreshing Materialized Views...');
   await sequelize.query(`REFRESH MATERIALIZED VIEW "TransactionBalances"`);
   await sequelize.query(`REFRESH MATERIALIZED VIEW "CollectiveBalanceCheckpoint"`);
   await sequelize.query(`REFRESH MATERIALIZED VIEW "CollectiveTransactionStats"`);
@@ -118,7 +113,7 @@ program.command('restore <file>').action(async file => {
   await sequelize.query(`REFRESH MATERIALIZED VIEW "ExpenseTagStats"`);
   await sequelize.query(`REFRESH MATERIALIZED VIEW "HostMonthlyTransactions"`);
 
-  console.log('\n>>> Done!');
+  logger.info('>>> Done!');
   sequelize.close();
 });
